+++ game_patch/graphics/d3d11/gr_d3d11_hooks.h
@@ -9,5 +9,6 @@ namespace df::gr::d3d11 {
                       float sx, float sy, float sw, float sh,
                       bool flip_x, bool flip_y, rf::gr::Mode mode);
     void update_texture_filtering();
+    void set_pow2_tex_active(bool active);
 }